refactor: Move main.go to root directory
