+++ .air.toml
@@ -5,7 +5,7 @@ tmp_dir = "tmp"
 [build]
   args_bin = []
   bin = "./bin/server"
-  cmd = "go build -o ./bin/server ./cmd/server/main.go"
+  cmd = "go build -o ./bin/server ./main.go"
   delay = 1000
   exclude_dir = ["assets", "tmp", "vendor", "testdata"]
   exclude_file = []
