+++ Dockerfile
@@ -10,9 +10,9 @@ RUN go install github.com/swaggo/swag/v2/cmd/swag@v2.0.0-rc4
 
 COPY . .
 
-RUN	swag init -v3.1 -o docs -g cmd/server/main.go
+RUN	swag init -v3.1 -o docs -g main.go
 
-RUN go build -o bin/server ./cmd/server/main.go
+RUN go build -o bin/server ./main.go
 
 RUN touch /app/.env
