+++ Makefile
@@ -1,8 +1,8 @@
 build:
-	go build -o bin/server ./cmd/server/main.go
+	go build -o bin/server ./main.go
 
-run: 
-	go run ./cmd/server/main.go
+run:
+	go run ./main.go
 
 dev:
 	air
@@ -17,7 +17,7 @@ deps:
 	go mod tidy
 
 gen-docs:
-	swag init -v3.1 -o docs -g cmd/server/main.go
+	swag init -v3.1 -o docs -g main.go
 
 lint:
 	golangci-lint run
